refactor: use afero, use inmem filesystem for wasm, implement some datastore methods
