+++ cmd/wasmlight/light.go
@@ -7,8 +7,10 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"strings"
 	"syscall/js"
 
+	"github.com/BurntSushi/toml"
 	"github.com/celestiaorg/celestia-app/app"
 	"github.com/celestiaorg/celestia-app/app/encoding"
 	"github.com/celestiaorg/celestia-node/nodebuilder"
@@ -37,18 +39,20 @@ func main() {
 		appendLog.Invoke(msg, level)
 	}
 
-	ctx, stop := context.WithCancel(context.Background())
-	cancel = stop
+	var ctx context.Context
+	ctx, cancel = context.WithCancel(context.Background())
 
 	js.Global().Set("startNode", js.FuncOf(func(this js.Value, args []js.Value) interface{} {
+		configStr := args[0].String()
+
 		fmt.Println("attempting to start node")
-		go Start(ctx, log)
+		go start(ctx, log, configStr)
 		return nil
 	}))
 
 	js.Global().Set("stopNode", js.FuncOf(func(this js.Value, args []js.Value) interface{} {
 		fmt.Println("attempting to stop node")
-		go Stop(ctx, log)
+		go stop(ctx, log)
 		return nil
 	}))
 
@@ -59,47 +63,52 @@ func main() {
 	}
 }
 
-func Start(ctx context.Context, log func(msg string, level string)) error {
+func start(ctx context.Context, log func(msg string, level string), configStr string) {
+	cfg := &nodebuilder.Config{}
+	toml.Decode(configStr, cfg)
+
 	encConf := encoding.MakeConfig(app.ModuleEncodingRegisters...)
 	ring, err := keyring.New(app.Name, keyring.BackendMemory, "", os.Stdin, encConf.Codec)
 	if err != nil {
 		log(fmt.Sprintf("Failed to create keyring: %s", err), "error")
-		return err
+		return
 	}
 
-	store, err := nodebuilder.OpenStore(".celestia-light-arabica-10", ring)
+	basePath := ".celestia-light-arabica-10"
+	if err := nodebuilder.SaveConfig(strings.Join([]string{basePath, "config.toml"}, "/"), cfg); err != nil {
+		log(fmt.Sprintf("unable to save config %s", err), "error")
+		return
+	}
+
+	store, err := nodebuilder.OpenStore(basePath, ring)
 	if err != nil {
 		log(fmt.Sprintf("Failed to open store: %s", err), "error")
-		return err
+		return
 	}
 	defer store.Close()
 
 	nd, err = nodebuilder.New(node.Light, p2p.Arabica, store)
 	if err != nil {
 		log(fmt.Sprintf("Failed to create new node: %s", err), "error")
-		return err
+		return
 	}
-
 	if err := nd.Start(ctx); err != nil {
 		log(fmt.Sprintf("Failed to start node: %s", err), "error")
-		return err
+		return
 	}
 
 	log("Node started successfully", "info")
 	<-ctx.Done()
-	return nil
+	return
 }
 
-func Stop(ctx context.Context, log func(msg string, level string)) error {
-	if nd != nil {
-		if err := nd.Stop(ctx); err != nil {
-			log(fmt.Sprintf("Failed to stop node: %s", err), "error")
-			return err
-		}
-		log("Node stopped successfully", "info")
-	} else {
-		log("No node instance found to stop", "warn")
+func stop(ctx context.Context, log func(msg string, level string)) error {
+	if err := nd.Stop(ctx); err != nil {
+		log(fmt.Sprintf("Failed to stop node: %s", err), "error")
+		return err
 	}
+
+	log("Node stopped successfully", "info")
 	if cancel != nil {
 		cancel()
 	}
