+++ libs/keystore/fs_keystore.go
@@ -1,11 +1,11 @@
-//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris || zos
-
 package keystore
 
 import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/spf13/afero"
+	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
@@ -19,27 +19,29 @@ var ErrNotFound = errors.New("keystore: key not found")
 // fsKeystore implements persistent Keystore over OS filesystem.
 type fsKeystore struct {
 	path string
+	fs   afero.Fs
 
 	ring keyring.Keyring
 }
 
 // NewFSKeystore creates a new Keystore over OS filesystem.
 // The path must point to a directory. It is created automatically if necessary.
-func NewFSKeystore(path string, ring keyring.Keyring) (Keystore, error) {
-	err := os.Mkdir(path, 0755)
+func NewFSKeystore(path string, ring keyring.Keyring, fs afero.Fs) (Keystore, error) {
+	err := fs.Mkdir(path, 0755)
 	if err != nil && !os.IsExist(err) {
 		return nil, fmt.Errorf("keystore: failed to make a dir: %w", err)
 	}
 	return &fsKeystore{
 		path: path,
 		ring: ring,
+		fs:   fs,
 	}, nil
 }
 
 func (f *fsKeystore) Put(n KeyName, pk PrivKey) error {
 	path := f.pathTo(n.Base32())
 
-	_, err := os.Stat(path)
+	_, err := f.fs.Stat(path)
 	if err == nil {
 		return fmt.Errorf("keystore: key '%s' already exists", n)
 	} else if !os.IsNotExist(err) {
@@ -52,6 +54,14 @@ func (f *fsKeystore) Put(n KeyName, pk PrivKey) error {
 	}
 
 	err = os.WriteFile(path, data, 0600)
+	file, err := f.fs.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	if _, err := file.Write(data); err != nil {
+		return err
+	}
 	if err != nil {
 		return fmt.Errorf("keystore: failed to write key '%s': %w", n, err)
 	}
@@ -61,7 +71,7 @@ func (f *fsKeystore) Put(n KeyName, pk PrivKey) error {
 func (f *fsKeystore) Get(n KeyName) (PrivKey, error) {
 	path := f.pathTo(n.Base32())
 
-	st, err := os.Stat(path)
+	st, err := f.fs.Stat(path)
 	if err != nil {
 		if os.IsNotExist(err) {
 			return PrivKey{}, fmt.Errorf("%w: %s", ErrNotFound, n)
@@ -74,7 +84,12 @@ func (f *fsKeystore) Get(n KeyName) (PrivKey, error) {
 		return PrivKey{}, fmt.Errorf("keystore: permissions of key '%s' are too relaxed: %w", n, err)
 	}
 
-	data, err := os.ReadFile(path)
+	file, err := f.fs.Open(path)
+	if err != nil {
+		return PrivKey{}, fmt.Errorf("keystore: failed to open file key '%s': %w", n, err)
+	}
+	defer file.Close()
+	data, err := io.ReadAll(file)
 	if err != nil {
 		return PrivKey{}, fmt.Errorf("keystore: failed read key '%s': %w", n, err)
 	}
@@ -90,20 +105,21 @@ func (f *fsKeystore) Get(n KeyName) (PrivKey, error) {
 func (f *fsKeystore) Delete(n KeyName) error {
 	path := f.pathTo(n.Base32())
 
-	_, err := os.Stat(path)
+	_, err := f.fs.Stat(path)
 	if os.IsNotExist(err) {
 		return fmt.Errorf("keystore: key '%s' not found", n)
 	} else if err != nil {
 		return fmt.Errorf("keystore: check before reading key '%s' failed: %w", n, err)
 	}
 
-	err = os.Remove(path)
+	err = f.fs.Remove(path)
 	if err != nil {
 		return fmt.Errorf("keystore: failed to delete key '%s': %w", n, err)
 	}
 	return nil
 }
 
+// FIXME implement using afero for compatibility
 func (f *fsKeystore) List() ([]KeyName, error) {
 	entries, err := fs.ReadDir(os.DirFS(filepath.Dir(f.path)), filepath.Base(f.path))
 	if err != nil {
