+++ nodebuilder/fs_wasm.go
@@ -1,0 +1,7 @@
+//go:build wasm
+
+package nodebuilder
+
+import "github.com/spf13/afero"
+
+var fs = afero.NewMemMapFs()
